backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "chromadb>=1.5.8",
    "fastapi>=0.136.0",
    "firebase-admin>=7.4.0",
    "google-genai>=1.73.1",
    "llama-index>=0.14.20",
    "llama-index-embeddings-google-genai>=0.5.0",
    "llama-index-llms-google-genai>=0.9.1",
    "llama-index-readers-file>=0.6.0",
    "llama-index-vector-stores-chroma>=0.5.5",
    "markdown>=3.10.2",
    "pillow>=12.2.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "python-pptx>=1.0.2",
    "tree-sitter>=0.25.2",
    "tree-sitter-languages>=1.10.2",
    "uvicorn>=0.44.0",
    "xhtml2pdf>=0.2.17",
]
